requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.2.0
python-decouple==3.8

# Database dependencies
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Task queue and scheduling
apscheduler==3.10.4

# Email service
sendgrid==6.10.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging and monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Rate limiting
slowapi==0.1.9

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Additional dependencies
watchfiles
python-slugify 